backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
docxtpl
python-docx
